+++ lbfgs/linesearch_gpu.h
@@ -54,8 +54,8 @@ namespace gpu_lbfgs
 	__global__ void strongWolfePhase2(size_t tries);
 }
 
-bool lbfgs::gpu_linesearch(float *d_x, float *d_z, float *d_fk, float *d_gk,
-						   size_t &evals, const float *d_gkm1, float *d_fkm1,
+bool lbfgs::gpu_linesearch(float *d_x, float *d_z, double *d_fk, float *d_gk,
+						   size_t &evals, const float *d_gkm1, double *d_fkm1,
 						   lbfgs::status &stat, float *step, size_t maxEvals,
 						   timer *timer_evals, timer *timer_linesearch,
 						   float *d_tmp, int *d_status)
@@ -67,8 +67,8 @@ bool lbfgs::gpu_linesearch(float *d_x, float *d_z, float *d_fk, float *d_gk,
 	CudaSafeCall( cudaMalloc((void**)&d_x_original, NX * sizeof(float)) );
 	CudaSafeCall( cudaMemcpy(d_x_original, d_x, NX * sizeof(float), cudaMemcpyDeviceToDevice) );
 
-	float F = 0.0f, F_original = 0.0f;
-	CudaSafeCall( cudaMemcpy(&F_original, d_fk, sizeof(float), cudaMemcpyDeviceToHost) );
+	double F = 0.0f, F_original = 0.0f;
+	CudaSafeCall( cudaMemcpy(&F_original, d_fk, sizeof(double), cudaMemcpyDeviceToHost) );
 
 	float phi_prime_0;
 
@@ -166,10 +166,10 @@ bool lbfgs::gpu_linesearch(float *d_x, float *d_z, float *d_fk, float *d_gk,
 		// If both Armijo and Strong Wolfe hold, we're done
 		if (ret == 1)
 		{
-			CudaSafeCall( cudaMemcpy(&F, d_fk, sizeof(float), cudaMemcpyDeviceToHost) );
+			CudaSafeCall( cudaMemcpy(&F, d_fk, sizeof(double), cudaMemcpyDeviceToHost) );
 			if(F > F_original){
 				CudaSafeCall( cudaMemcpy(d_x, d_x_original, NX * sizeof(float), cudaMemcpyDeviceToDevice) );
-				CudaSafeCall( cudaMemcpy(d_fk, &F_original, sizeof(float), cudaMemcpyHostToDevice) );
+				CudaSafeCall( cudaMemcpy(d_fk, &F_original, sizeof(double), cudaMemcpyHostToDevice) );
 			}
 			CudaSafeCall( cudaFree(d_x_original) );
 			return true;
@@ -195,10 +195,10 @@ bool lbfgs::gpu_linesearch(float *d_x, float *d_z, float *d_fk, float *d_gk,
 		if (ret == 3)
 		{
 			stat = lbfgs::LBFGS_LINE_SEARCH_FAILED;
-			CudaSafeCall( cudaMemcpy(&F, d_fk, sizeof(float), cudaMemcpyDeviceToHost) );
+			CudaSafeCall( cudaMemcpy(&F, d_fk, sizeof(double), cudaMemcpyDeviceToHost) );
 			if(F > F_original){
 				CudaSafeCall( cudaMemcpy(d_x, d_x_original, NX * sizeof(float), cudaMemcpyDeviceToDevice) );
-				CudaSafeCall( cudaMemcpy(d_fk, &F_original, sizeof(float), cudaMemcpyHostToDevice) );
+				CudaSafeCall( cudaMemcpy(d_fk, &F_original, sizeof(double), cudaMemcpyHostToDevice) );
 			}
 			CudaSafeCall( cudaFree(d_x_original) );
 			return false;
@@ -260,10 +260,10 @@ bool lbfgs::gpu_linesearch(float *d_x, float *d_z, float *d_fk, float *d_gk,
 		if (ret == 1)
 		{
 			// The Armijo and Strong Wolfe conditions hold
-			CudaSafeCall( cudaMemcpy(&F, d_fk, sizeof(float), cudaMemcpyDeviceToHost) );
+			CudaSafeCall( cudaMemcpy(&F, d_fk, sizeof(double), cudaMemcpyDeviceToHost) );
 			if(F > F_original){
 				CudaSafeCall( cudaMemcpy(d_x, d_x_original, NX * sizeof(float), cudaMemcpyDeviceToDevice) );
-				CudaSafeCall( cudaMemcpy(d_fk, &F_original, sizeof(float), cudaMemcpyHostToDevice) );
+				CudaSafeCall( cudaMemcpy(d_fk, &F_original, sizeof(double), cudaMemcpyHostToDevice) );
 			}
 			CudaSafeCall( cudaFree(d_x_original) );
 			return true;
@@ -273,10 +273,10 @@ bool lbfgs::gpu_linesearch(float *d_x, float *d_z, float *d_fk, float *d_gk,
 		{
 			// The search interval has become too small
 			stat = lbfgs::LBFGS_LINE_SEARCH_FAILED;
-			CudaSafeCall( cudaMemcpy(&F, d_fk, sizeof(float), cudaMemcpyDeviceToHost) );
+			CudaSafeCall( cudaMemcpy(&F, d_fk, sizeof(double), cudaMemcpyDeviceToHost) );
 			if(F > F_original){
 				CudaSafeCall( cudaMemcpy(d_x, d_x_original, NX * sizeof(float), cudaMemcpyDeviceToDevice) );
-				CudaSafeCall( cudaMemcpy(d_fk, &F_original, sizeof(float), cudaMemcpyHostToDevice) );
+				CudaSafeCall( cudaMemcpy(d_fk, &F_original, sizeof(double), cudaMemcpyHostToDevice) );
 			}
 			CudaSafeCall( cudaFree(d_x_original) );
 			return false;
@@ -285,10 +285,10 @@ bool lbfgs::gpu_linesearch(float *d_x, float *d_z, float *d_fk, float *d_gk,
 		if (evals >= maxEvals)
 		{
 			stat = lbfgs::LBFGS_REACHED_MAX_EVALS;	
-			CudaSafeCall( cudaMemcpy(&F, d_fk, sizeof(float), cudaMemcpyDeviceToHost) );
+			CudaSafeCall( cudaMemcpy(&F, d_fk, sizeof(double), cudaMemcpyDeviceToHost) );
 			if(F > F_original){
 				CudaSafeCall( cudaMemcpy(d_x, d_x_original, NX * sizeof(float), cudaMemcpyDeviceToDevice) );
-				CudaSafeCall( cudaMemcpy(d_fk, &F_original, sizeof(float), cudaMemcpyHostToDevice) );
+				CudaSafeCall( cudaMemcpy(d_fk, &F_original, sizeof(double), cudaMemcpyHostToDevice) );
 			}
 			CudaSafeCall( cudaFree(d_x_original) );
 			return false;
